move netdataapi pkg
