+++ src/go/plugin/go.d/agent/module/job.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	"github.com/netdata/netdata/go/plugins/logger"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/netdataapi"
+	"github.com/netdata/netdata/go/plugins/pkg/netdataapi"
 )
 
 var obsoleteLock = &sync.Mutex{}
